+++ masp_primitives/src/group_hash.rs
@@ -2,7 +2,7 @@
 //!
 //! [grouphash]: https://zips.z.cash/protocol/protocol.pdf#concretegrouphashjubjub
 
-use group::{cofactor::CofactorGroup, Group, GroupEncoding, ff::PrimeField};
+use group::{cofactor::CofactorGroup, ff::PrimeField, Group, GroupEncoding};
 
 use crate::constants;
 use blake2s_simd::Params;
